requirements.txt
beautifulsoup4==4.6.0
Django==4.0
djangorestframework==3.12.4
googlemaps==4.5.3
icecream==2.1.1

pip~=21.0.1
zlib~=1.2.11
wheel~=0.36.2
brotli~=1.0.9
openssl~=1.1.1k
cryptography~=3.4.7
py~=1.10.0
lxml~=4.6.3
keyring~=22.3.0
typing_extensions~=3.7.4.3
mypy_extensions~=0.4.1
setuptools~=52.0.0
cython~=0.29.23
docutils~=0.17
future~=0.18.2
pandas~=1.2.4
requests~=2.25.1
distributed~=2021.4.0
scrapy~=2.5.1
mysql-connector-python~=8.0.27